+++ CHANGELOG.md
@@ -24,6 +24,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - For plugin < v4.6.0 use Homebridge UI <= v5.5.0
 - For plugin >= v4.6.0 use Homebridge UI >= v5.13.0
 
+# [4.8.4] - (22.03.2026)
+
+## Changes
+
+- fix report correct fan mode fter power ON with HK
+- fix wide vane support
+- cleanup payload before send
+- bump dependencies
+- cleanup
+
 # [4.8.3] - (17.03.2026)
 
 ## Changes